backend final

+++ src/main/java/com/com/dualexpress/dto/LigneCommandeDTO.java
@@ -1,7 +1,8 @@
-
 package com.dualexpress.dto;
 
 import lombok.*;
+
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -11,6 +12,6 @@ public class LigneCommandeDTO {
 
     private Long id;
     private Integer quantite;
-    private Double prixUnitaire;
+    private BigDecimal prixUnitaire;
     private Long produitId;
 }
