+++ src/main/java/com/com/dualexpress/dto/ProduitDTO.java
@@ -1,18 +1,18 @@
-
 package com.dualexpress.dto;
 
 import lombok.*;
+
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class ProduitDTO {
-
     private Long id;
     private String nom;
     private String description;
-    private Double prix;
+    private BigDecimal prix;
     private String categorie;
     private Boolean disponible;
     private Long restaurantId;
